+++ packages/cli/__tests__/daemon/__integration__/disable/__scripts__/daemon_disable.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+
+echo "================ daemon disable ==============="
+zowe daemon disable
+exit $?