+++ packages/cli/__tests__/daemon/__integration__/enable/__scripts__/daemon_enable.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+set -e
+
+echo "================ daemon enable ==============="
+zowe daemon enable
+exit $?