+++ packages/cli/__tests__/daemon/__integration__/enable/__scripts__/daemon_enable_help.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+set -e
+
+echo "================ daemon enable help ==============="
+zowe daemon enable --help
+exit $?
+